Move pydantic models to their own file
